Bug fixes.
Add ability to process multiple packets in a buffer.  This includes the
ability to deal with packets split amongst multiple incoming buffers,
but so far I haven't been able to produce this situation in testing.
Its possible that a buffer is only accumulates as long as a full
transfer fits in the buffer of 16384 bytes